Bump to v3.0.1 and reimplement CustomFileServerList better

+++ src/main/java/xyz/mrmelon54/MultipleServerLists/util/CustomFileServerList.java
@@ -1,6 +1,5 @@
 package xyz.mrmelon54.MultipleServerLists.util;
 
-import com.google.common.collect.Lists;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
 import net.fabricmc.fabric.api.util.NbtType;
@@ -11,8 +10,10 @@ import net.minecraft.nbt.NbtCompound;
 import net.minecraft.nbt.NbtIo;
 import net.minecraft.nbt.NbtList;
 import net.minecraft.util.Util;
+import org.apache.commons.compress.utils.Lists;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import xyz.mrmelon54.MultipleServerLists.duck.ServerListDuckProvider;
 
 import java.io.File;
 import java.util.List;
@@ -22,7 +23,6 @@ public class CustomFileServerList extends ServerList {
     private static final Logger LOGGER = LogManager.getLogger();
     private File serversWrapperFolder;
     private String listName;
-    private List<ServerInfo> customServers;
     private final int pageIndex;
 
     public CustomFileServerList(MinecraftClient client, int pageIndex) {
@@ -32,15 +32,29 @@ public class CustomFileServerList extends ServerList {
         this.loadFile();
     }
 
+    List<ServerInfo> getInternalServers() {
+        if (this instanceof ServerListDuckProvider duck) return duck.getServers();
+        return Lists.newArrayList();
+    }
+
+    List<ServerInfo> getInternalHiddenServers() {
+        if (this instanceof ServerListDuckProvider duck) return duck.getHiddenServers();
+        return Lists.newArrayList();
+    }
+
+    public boolean makeSureFolderExists() {
+        if (this.serversWrapperFolder == null)
+            this.serversWrapperFolder = new File(MinecraftClient.getInstance().runDirectory, "s760");
+        return serversWrapperFolder.exists() || serversWrapperFolder.mkdirs();
+    }
+
     @Override
     public void loadFile() {
         try {
-            if (this.customServers == null)
-                this.customServers = Lists.newArrayList();
-            this.customServers.clear();
-            if (this.serversWrapperFolder == null)
-                this.serversWrapperFolder = new File(MinecraftClient.getInstance().runDirectory, "s760");
-            if (serversWrapperFolder.exists() || serversWrapperFolder.mkdirs()) {
+            getInternalServers().clear();
+            getInternalHiddenServers().clear();
+
+            if (makeSureFolderExists()) {
                 NbtCompound nbtCompound = NbtIo.read(new File(serversWrapperFolder, "servers" + pageIndex + ".dat"));
                 if (nbtCompound == null)
                     return;
@@ -49,8 +63,12 @@ public class CustomFileServerList extends ServerList {
                     listName = nbtCompound.getString("name");
                 NbtList nbtList = nbtCompound.getList("servers", NbtType.COMPOUND);
 
-                for (int i = 0; i < nbtList.size(); ++i)
-                    this.customServers.add(ServerInfo.fromNbt(nbtList.getCompound(i)));
+                for (int i = 0; i < nbtList.size(); ++i) {
+                    NbtCompound c = nbtList.getCompound(i);
+                    ServerInfo serverInfo = ServerInfo.fromNbt(c);
+                    if (c.getBoolean("hidden")) getInternalHiddenServers().add(serverInfo);
+                    else getInternalServers().add(serverInfo);
+                }
             }
         } catch (Exception var4) {
             LOGGER.error("Couldn't load server list", var4);
@@ -60,22 +78,30 @@ public class CustomFileServerList extends ServerList {
     @Override
     public void saveFile() {
         try {
-            if (this.customServers == null)
-                this.customServers = Lists.newArrayList();
-            if (this.serversWrapperFolder == null)
-                this.serversWrapperFolder = new File(MinecraftClient.getInstance().runDirectory, "s760");
+            if (makeSureFolderExists()) {
+                NbtList nbtList = new NbtList();
+                for (ServerInfo serverInfo : getInternalServers()) {
+                    NbtCompound c = serverInfo.toNbt();
+                    c.putBoolean("hidden", false);
+                    nbtList.add(c);
+                }
+                for (ServerInfo serverInfo : getInternalHiddenServers()) {
+                    NbtCompound c = serverInfo.toNbt();
+                    c.putBoolean("hidden", true);
+                    nbtList.add(c);
+                }
 
-            NbtList nbtList = new NbtList();
-            for (ServerInfo serverInfo : this.customServers) nbtList.add(serverInfo.toNbt());
+                NbtCompound nbtCompound = new NbtCompound();
+                nbtCompound.putString("name", listName);
+                nbtCompound.put("servers", nbtList);
 
-            NbtCompound nbtCompound = new NbtCompound();
-            nbtCompound.putString("name", listName);
-            nbtCompound.put("servers", nbtList);
-            File file = File.createTempFile("servers" + pageIndex, ".dat", this.serversWrapperFolder);
-            NbtIo.write(nbtCompound, file);
-            File file2 = new File(this.serversWrapperFolder, "servers" + pageIndex + ".dat_old");
-            File file3 = new File(this.serversWrapperFolder, "servers" + pageIndex + ".dat");
-            Util.backupAndReplace(file3, file, file2);
+                String n = "servers" + pageIndex;
+                File file = File.createTempFile(n, ".dat", this.serversWrapperFolder);
+                NbtIo.write(nbtCompound, file);
+                File file2 = new File(this.serversWrapperFolder, n + ".dat_old");
+                File file3 = new File(this.serversWrapperFolder, n + ".dat");
+                Util.backupAndReplace(file3, file, file2);
+            }
         } catch (Exception var6) {
             LOGGER.error("Couldn't save server list", var6);
         }
@@ -83,54 +109,20 @@ public class CustomFileServerList extends ServerList {
 
     public void deleteFile() {
         try {
-            if (this.customServers == null)
-                this.customServers = Lists.newArrayList();
-            if (this.serversWrapperFolder == null)
-                this.serversWrapperFolder = new File(MinecraftClient.getInstance().runDirectory, "s760");
-
-            File file3 = new File(this.serversWrapperFolder, "servers" + pageIndex + ".dat");
-            file3.delete();
+            if (makeSureFolderExists()) {
+                File file3 = new File(this.serversWrapperFolder, "servers" + pageIndex + ".dat");
+                file3.delete();
+            }
         } catch (Exception var6) {
             LOGGER.error("Couldn't remove the server list", var6);
         }
     }
 
-    public ServerInfo get(int index) {
-        return this.customServers.get(index);
-    }
-
-    public void remove(ServerInfo serverInfo) {
-        this.customServers.remove(serverInfo);
-    }
-
-    public void add(ServerInfo serverInfo) {
-        this.customServers.add(serverInfo);
-    }
-
-    public void add(int index, ServerInfo serverInfo) {
-        this.customServers.add(index, serverInfo);
-    }
-
-    public int size() {
-        return this.customServers.size();
-    }
-
-    public void swapEntries(int index1, int index2) {
-        ServerInfo serverInfo = this.get(index1);
-        this.customServers.set(index1, this.get(index2));
-        this.customServers.set(index2, serverInfo);
-        this.saveFile();
-    }
-
-    public void set(int index, ServerInfo serverInfo) {
-        this.customServers.set(index, serverInfo);
+    public void setName(String value) {
+        listName = value;
     }
 
     public String getName() {
         return listName;
-    }
-
-    public void setName(String value) {
-        listName = value;
     }
 }
